feat(core): add PathHelper.GetUniquePath with tar.gz support
Extract tar.gz-aware GetUniquePath from App.xaml.cs into Core/Utils/PathHelper.cs. FileConflictHelper.GetUniquePath and App.GetUniquePath delegate to PathHelper to eliminate duplicate logic.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
