+++ src/MantisZip.UI/App.xaml.cs
@@ -358,34 +358,12 @@ public partial class App : Application
     /// </summary>
     internal static string GetUniquePath(string path)
     {
-        var dir = Path.GetDirectoryName(path) ?? ".";
-        string name, ext;
-
-        // 特别处理 .tar.gz 双扩展名
-        if (path.EndsWith(".tar.gz", StringComparison.OrdinalIgnoreCase))
-        {
-            // "C:\path\archive.tar.gz" → name="archive", ext=".tar.gz"
-            var withoutExt = path[..^7]; // 去掉 .tar.gz
-            name = Path.GetFileName(withoutExt);
-            ext = ".tar.gz";
-        }
+        var result = PathHelper.GetUniquePath(path);
+        if (result != path)
+            LogDebug("GetUniquePath: {0} → {1}", path, result);
         else
-        {
-            name = Path.GetFileNameWithoutExtension(path);
-            ext = Path.GetExtension(path);
-        }
-
-        for (int i = 1; i < 1000; i++)
-        {
-            var candidate = Path.Combine(dir, $"{name} ({i}){ext}");
-            if (!File.Exists(candidate))
-            {
-                LogDebug("GetUniquePath: {0} → {1}", path, candidate);
-                return candidate;
-            }
-        }
-        LogDebug("GetUniquePath: {0} → {0} (all names taken, fallback)", path);
-        return path; // 999 个名字全被占用了，直接覆盖原文件
+            LogDebug("GetUniquePath: {0} → {0} (all names taken, fallback)", path);
+        return result;
     }
 
     internal static FileConflictAction GetConflictActionFromSettings()
